implications of stricter type equality: if both types are named, they must be the same type (arising from the same declaration).
R=r,gri
DELTA=44  (21 added, 4 deleted, 19 changed)
OCL=28436
CL=28577

+++ test/interface10.go
@@ -13,7 +13,7 @@ const Value = 1e12
 type Inter interface { M() int64 }
 
 type T int64
-func (t T) M() int64 { return t }
+func (t T) M() int64 { return int64(t) }
 var t = T(Value)
 var pt = &t
 var ti Inter = t
